refactor: extract steps registry
