+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/steps/stepsRegistry.tsx
@@ -1,0 +1,107 @@
+import React, { ComponentType } from 'react'
+
+import { CancelledStep } from './CancelledStep'
+import { CancellingStep } from './CancellingStep'
+import { ExecutingStep } from './ExecutingStep'
+import { ExpiredStep } from './ExpiredStep'
+import { FinishedStep } from './FinishedStep'
+import { InitialStep } from './InitialStep'
+import { SolvingStep } from './SolvingStep'
+
+import { OrderProgressBarProps, OrderProgressBarStepName } from '../../types'
+import { RenderProgressTopSection } from '../RenderProgressTopSection'
+
+const DEBUG_FORCE_SHOW_SURPLUS = false
+
+function InitialStepWrapper(props: OrderProgressBarProps) {
+  return (
+    <InitialStep>
+      <RenderProgressTopSection {...props} debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS} />
+    </InitialStep>
+  )
+}
+
+function ExecutingStepWrapper(props: OrderProgressBarProps) {
+  return (
+    <ExecutingStep>
+      <RenderProgressTopSection {...props} debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS} />
+    </ExecutingStep>
+  )
+}
+
+function FinishedStepWrapper(props: OrderProgressBarProps) {
+  const { stepName, solverCompetition: solvers, totalSolvers, order, surplusData, chainId, receiverEnsName } = props
+
+  return (
+    <FinishedStep
+      stepName={stepName}
+      surplusData={surplusData}
+      solvers={solvers}
+      order={order}
+      chainId={chainId}
+      receiverEnsName={receiverEnsName}
+      totalSolvers={totalSolvers}
+      debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS}
+    >
+      <RenderProgressTopSection {...props} debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS} />
+    </FinishedStep>
+  )
+}
+
+function SolvingStepWrapper(props: OrderProgressBarProps) {
+  const { countdown, stepName, showCancellationModal } = props
+  const isUnfillable = stepName === 'unfillable'
+  const isDelayed = stepName === 'delayed'
+  const isSubmissionFailed = stepName === 'submissionFailed'
+  const isSolved = stepName === 'solved'
+  const calculatedCountdownValue = isUnfillable || isDelayed || isSubmissionFailed || isSolved ? undefined : countdown
+
+  return (
+    <SolvingStep stepName={stepName} showCancellationModal={showCancellationModal}>
+      <RenderProgressTopSection
+        {...props}
+        countdown={calculatedCountdownValue}
+        debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS}
+      />
+    </SolvingStep>
+  )
+}
+
+function CancellingStepWrapper(props: OrderProgressBarProps) {
+  return (
+    <CancellingStep>
+      <RenderProgressTopSection {...props} debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS} />
+    </CancellingStep>
+  )
+}
+
+function CancelledStepWrapper(props: OrderProgressBarProps) {
+  return (
+    <CancelledStep>
+      <RenderProgressTopSection {...props} debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS} />
+    </CancelledStep>
+  )
+}
+
+function ExpiredStepWrapper(props: OrderProgressBarProps) {
+  return (
+    <ExpiredStep navigateToNewOrder={props.navigateToNewOrder}>
+      <RenderProgressTopSection {...props} debugForceShowSurplus={DEBUG_FORCE_SHOW_SURPLUS} />
+    </ExpiredStep>
+  )
+}
+
+export const STEP_NAME_TO_STEP_COMPONENT: Record<OrderProgressBarStepName, ComponentType<OrderProgressBarProps>> = {
+  initial: InitialStepWrapper,
+  solving: SolvingStepWrapper,
+  executing: ExecutingStepWrapper,
+  finished: FinishedStepWrapper,
+  solved: SolvingStepWrapper, // Use SolvingStep for 'solved' state
+  delayed: SolvingStepWrapper,
+  unfillable: SolvingStepWrapper,
+  submissionFailed: SolvingStepWrapper,
+  cancelling: CancellingStepWrapper,
+  cancelled: CancelledStepWrapper,
+  expired: ExpiredStepWrapper,
+  cancellationFailed: FinishedStepWrapper,
+}
